backgroundMellow/backend/requirements.txt
### Core web framework
fastapi==0.121.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
starlette==0.49.3

### Data validation / settings
pydantic==2.11.10
pydantic-core==2.33.2

### Audio processing & evaluation
pydub==0.25.1
librosa==0.9.2
soundfile==0.12.1
laion-clap==1.1.7
frechet_audio_distance==0.3.3
pyloudnorm==0.2.0
pystoi==0.4.1
torchlibrosa==0.1.0

### Core ML / scientific stack (aligned to your working env)
torch==2.4.1
torchaudio>=2.0.0

### Hugging Face / diffusion ecosystem
transformers==4.46.1
accelerate==0.18.0
diffusers==0.18.2
einops==0.6.1
huggingface-hub==0.33.4
resampy==0.4.2
tqdm==4.63.1

### Audio generation model wrappers
# tango==0.2.2
# tangoflux==0.1.0
# parler_tts==0.2.2
# elevenlabs==2.38.1

### NLP / text stack
spacy==3.8.11
spacy-curated-transformers==0.3.1

### LLM / orchestration
openai==2.26.0
google-genai==1.57.0
google-generativeai==0.8.6
langchain==1.2.0
langchain-core==1.2.6
langchain-huggingface==1.2.0
langgraph==1.0.5
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0

### Config / utilities
python-dotenv==1.2.1
requests==2.32.5
protobuf>=4.25.8
setuptools==70.0.0
importlib-metadata==6.3.0
progressbar33==2.4
wandb==0.25.0
ipython==8.12.0
laion_clap

# panns_inference==0.1.3

elevenlabs==2.38.1

sentence_transformers==5.3.0

# install models from github
# TangoFlux must be installed from GitHub
# git+https://github.com/declare-lab/TangoFlux.git
git+https://github.com/huggingface/parler-tts.git

# pip3 install --pre torch torchaudio --index-url https://download.pytorch.org/whl/nightly/cpu